requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
marshmallow-sqlalchemy==1.4.2
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.7.1
flask-cors==6.0.2